move evaluation typeof calls to Identifiers and MemberExpressions to static evaluator
this removes a breaking change again
