+++ lib/APIPlugin.js
@@ -38,7 +38,7 @@ class APIPlugin {
 			params.normalModuleFactory.plugin("parser", parser => {
 				Object.keys(REPLACEMENTS).forEach(key => {
 					parser.plugin(`expression ${key}`, ParserHelpers.toConstantDependency(REPLACEMENTS[key]));
-					parser.evaluateTypeOfPlugin.plugin(`evaluate typeof ${key}`, ParserHelpers.evaluateToString(REPLACEMENT_TYPES[key]));
+					parser.plugin(`evaluate typeof ${key}`, ParserHelpers.evaluateToString(REPLACEMENT_TYPES[key]));
 				});
 				IGNORES.forEach(key => {
 					parser.plugin(key, ParserHelpers.skipTraversal);
