+++ lib/DefinePlugin.js
@@ -85,7 +85,7 @@ class DefinePlugin {
 						parser.plugin("expression " + key, ParserHelpers.toConstantDependency(code));
 					}
 					const typeofCode = isTypeof ? code : "typeof (" + code + ")";
-					parser.evaluateTypeOfPlugin.plugin("evaluate typeof " + key, (expr) => {
+					parser.plugin("evaluate typeof " + key, (expr) => {
 						/**
 						 * this is needed in case there is a recursion in the DefinePlugin
 						 * to prevent an endless recursion
@@ -112,7 +112,7 @@ class DefinePlugin {
 					const code = stringifyObj(obj);
 					parser.plugin("can-rename " + key, ParserHelpers.approve);
 					parser.plugin("evaluate Identifier " + key, (expr) => new BasicEvaluatedExpression().setRange(expr.range));
-					parser.evaluateTypeOfPlugin.plugin("evaluate typeof " + key, ParserHelpers.evaluateToString("object"));
+					parser.plugin("evaluate typeof " + key, ParserHelpers.evaluateToString("object"));
 					parser.plugin("expression " + key, ParserHelpers.toConstantDependency(code));
 					parser.plugin("typeof " + key, ParserHelpers.toConstantDependency(JSON.stringify("object")));
 				}
