+++ lib/ExtendedAPIPlugin.js
@@ -37,7 +37,7 @@ class ExtendedAPIPlugin {
 			params.normalModuleFactory.plugin("parser", (parser, parserOptions) => {
 				Object.keys(REPLACEMENTS).forEach(key => {
 					parser.plugin(`expression ${key}`, ParserHelpers.toConstantDependency(REPLACEMENTS[key]));
-					parser.evaluateTypeOfPlugin.plugin("evaluate typeof " + key, ParserHelpers.evaluateToString(REPLACEMENT_TYPES[key]));
+					parser.plugin("evaluate typeof " + key, ParserHelpers.evaluateToString(REPLACEMENT_TYPES[key]));
 				});
 			});
 		});
