+++ lib/HotModuleReplacementPlugin.js
@@ -195,7 +195,7 @@ HotModuleReplacementPlugin.prototype.apply = function(compiler) {
 
 		params.normalModuleFactory.plugin("parser", function(parser, parserOptions) {
 			parser.plugin("expression __webpack_hash__", ParserHelpers.toConstantDependency("__webpack_require__.h()"));
-			parser.evaluateTypeOfPlugin.plugin("evaluate typeof __webpack_hash__", ParserHelpers.evaluateToString("string"));
+			parser.plugin("evaluate typeof __webpack_hash__", ParserHelpers.evaluateToString("string"));
 			parser.plugin("evaluate Identifier module.hot", function(expr) {
 				return ParserHelpers.evaluateToBoolean(!!parser.state.compilation.hotUpdateChunkTemplate)(expr);
 			});
