+++ lib/Parser.js
@@ -38,44 +38,10 @@ const PARSE_OPTIONS_SCRIPT = {
 	}
 };
 
-class EvaluateTypeOfPlugin extends Tapable {
-	constructor(parser) {
-		super();
-		parser.plugin("evaluate UnaryExpression", (expression) => {
-			if(expression.operator !== "typeof") {
-				return;
-			}
-
-			if(expression.argument.type === "Identifier") {
-				var name = parser.nameInCurrentScope(expression.argument.name);
-				if(parser.scope.definitions.indexOf(name) === -1) {
-					return this.applyPluginsBailResult1("evaluate typeof " + name, expression);
-				}
-			}
-			if(expression.argument.type === "MemberExpression") {
-				var expressionArgument = expression.argument;
-				var exprName = [];
-				while(expressionArgument.type === "MemberExpression" && !expressionArgument.computed) {
-					exprName.unshift(parser.nameInCurrentScope(expressionArgument.property.name));
-					expressionArgument = expressionArgument.object;
-				}
-				if(expressionArgument.type === "Identifier") {
-					exprName.unshift(parser.nameInCurrentScope(expressionArgument.name));
-					if(parser.scope.definitions.indexOf(name) === -1) {
-						exprName = exprName.join(".");
-						return this.applyPluginsBailResult1("evaluate typeof " + exprName, expression);
-					}
-				}
-			}
-		});
-	}
-}
-
 class Parser extends Tapable {
 	constructor(options) {
 		super();
 		this.options = options;
-		this.evaluateTypeOfPlugin = new EvaluateTypeOfPlugin(this);
 		this.initializeEvaluating();
 	}
 
@@ -85,7 +51,7 @@ class Parser extends Tapable {
 		this.plugin("evaluate LogicalExpression", boundStaticEvaluator);
 		this.plugin("evaluate BinaryExpression", boundStaticEvaluator);
 		this.plugin("evaluate UnaryExpression", boundStaticEvaluator);
-		this.evaluateTypeOfPlugin.plugin("evaluate typeof undefined", function(expr) {
+		this.plugin("evaluate typeof undefined", function(expr) {
 			return new BasicEvaluatedExpression().setString("undefined").setRange(expr.range);
 		});
 		this.plugin("evaluate CallExpression", boundStaticEvaluator);
