+++ lib/dependencies/AMDPlugin.js
@@ -80,12 +80,12 @@ class AMDPlugin {
 				setExpressionToModule("define", "!!webpack amd define");
 				parser.plugin("expression __webpack_amd_options__", () =>
 					parser.state.current.addVariable("__webpack_amd_options__", JSON.stringify(amdOptions)));
-				parser.evaluateTypeOfPlugin.plugin("evaluate typeof define.amd", ParserHelpers.evaluateToString(typeof amdOptions));
-				parser.evaluateTypeOfPlugin.plugin("evaluate typeof require.amd", ParserHelpers.evaluateToString(typeof amdOptions));
+				parser.plugin("evaluate typeof define.amd", ParserHelpers.evaluateToString(typeof amdOptions));
+				parser.plugin("evaluate typeof require.amd", ParserHelpers.evaluateToString(typeof amdOptions));
 				parser.plugin("evaluate Identifier define.amd", ParserHelpers.evaluateToBoolean(true));
 				parser.plugin("evaluate Identifier require.amd", ParserHelpers.evaluateToBoolean(true));
 				parser.plugin("typeof define", ParserHelpers.toConstantDependency(JSON.stringify("function")));
-				parser.evaluateTypeOfPlugin.plugin("evaluate typeof define", ParserHelpers.evaluateToString("function"));
+				parser.plugin("evaluate typeof define", ParserHelpers.evaluateToString("function"));
 				parser.plugin("can-rename define", ParserHelpers.approve);
 				parser.plugin("rename define", (expr) => {
 					const dep = new AMDRequireItemDependency("!!webpack amd define", expr.range);
@@ -95,7 +95,7 @@ class AMDPlugin {
 					return false;
 				});
 				parser.plugin("typeof require", ParserHelpers.toConstantDependency(JSON.stringify("function")));
-				parser.evaluateTypeOfPlugin.plugin("evaluate typeof require", ParserHelpers.evaluateToString("function"));
+				parser.plugin("evaluate typeof require", ParserHelpers.evaluateToString("function"));
 			});
 		});
 		compiler.plugin("after-resolvers", () => {
