+++ lib/dependencies/CommonJsPlugin.js
@@ -55,10 +55,10 @@ class CommonJsPlugin {
 				const requireExpressions = ["require", "require.resolve", "require.resolveWeak"];
 				for(let expression of requireExpressions) {
 					parser.plugin(`typeof ${expression}`, ParserHelpers.toConstantDependency("function"));
-					parser.evaluateTypeOfPlugin.plugin(`evaluate typeof ${expression}`, ParserHelpers.evaluateToString("function"));
+					parser.plugin(`evaluate typeof ${expression}`, ParserHelpers.evaluateToString("function"));
 				}
 
-				parser.evaluateTypeOfPlugin.plugin("evaluate typeof module", ParserHelpers.evaluateToString("object"));
+				parser.plugin("evaluate typeof module", ParserHelpers.evaluateToString("object"));
 				parser.plugin("assign require", (expr) => {
 					// to not leak to global "require", we need to define a local require here.
 					const dep = new ConstDependency("var require;", 0);
@@ -76,7 +76,7 @@ class CommonJsPlugin {
 					return false;
 				});
 				parser.plugin("typeof module", () => true);
-				parser.evaluateTypeOfPlugin.plugin("evaluate typeof exports", ParserHelpers.evaluateToString("object"));
+				parser.plugin("evaluate typeof exports", ParserHelpers.evaluateToString("object"));
 				parser.apply(
 					new CommonJsRequireDependencyParserPlugin(options),
 					new RequireResolveDependencyParserPlugin(options)
