+++ lib/dependencies/HarmonyDetectionParserPlugin.js
@@ -34,7 +34,7 @@ module.exports = class HarmonyDetectionParserPlugin {
 		});
 		var nonHarmonyIdentifiers = ["define", "exports"];
 		nonHarmonyIdentifiers.forEach(identifer => {
-			parser.evaluateTypeOfPlugin.plugin(`evaluate typeof ${identifer}`, nullInHarmony);
+			parser.plugin(`evaluate typeof ${identifer}`, nullInHarmony);
 			parser.plugin(`typeof ${identifer}`, skipInHarmony);
 			parser.plugin(`evaluate ${identifer}`, nullInHarmony);
 			parser.plugin(`expression ${identifer}`, skipInHarmony);
