+++ lib/dependencies/RequireEnsurePlugin.js
@@ -31,7 +31,7 @@ class RequireEnsurePlugin {
 					return;
 
 				parser.apply(new RequireEnsureDependenciesBlockParserPlugin());
-				parser.evaluateTypeOfPlugin.plugin("evaluate typeof require.ensure", ParserHelpers.evaluateToString("function"));
+				parser.plugin("evaluate typeof require.ensure", ParserHelpers.evaluateToString("function"));
 				parser.plugin("typeof require.ensure", ParserHelpers.toConstantDependency(JSON.stringify("function")));
 			});
 		});
