+++ lib/dependencies/RequireIncludePlugin.js
@@ -24,7 +24,7 @@ class RequireIncludePlugin {
 					return;
 
 				parser.apply(new RequireIncludeDependencyParserPlugin());
-				parser.evaluateTypeOfPlugin.plugin("evaluate typeof require.include", ParserHelpers.evaluateToString("function"));
+				parser.plugin("evaluate typeof require.include", ParserHelpers.evaluateToString("function"));
 				parser.plugin("typeof require.include", ParserHelpers.toConstantDependency(JSON.stringify("function")));
 			});
 		});
