+++ lib/dependencies/SystemPlugin.js
@@ -18,17 +18,17 @@ class SystemPlugin {
 					return;
 
 				function setNotSupported(name) {
-					parser.evaluateTypeOfPlugin.plugin("evaluate typeof " + name, ParserHelpers.evaluateToString("undefined"));
+					parser.plugin("evaluate typeof " + name, ParserHelpers.evaluateToString("undefined"));
 					parser.plugin("expression " + name,
 						ParserHelpers.expressionIsUnsupported(name + " is not supported by webpack.")
 					);
 				}
 
 				parser.plugin("typeof System", ParserHelpers.toConstantDependency(JSON.stringify("object")));
-				parser.evaluateTypeOfPlugin.plugin("evaluate typeof System", ParserHelpers.evaluateToString("object"));
+				parser.plugin("evaluate typeof System", ParserHelpers.evaluateToString("object"));
 
 				parser.plugin("typeof System.import", ParserHelpers.toConstantDependency(JSON.stringify("function")));
-				parser.evaluateTypeOfPlugin.plugin("evaluate typeof System.import", ParserHelpers.evaluateToString("function"));
+				parser.plugin("evaluate typeof System.import", ParserHelpers.evaluateToString("function"));
 
 				setNotSupported("System.set");
 				setNotSupported("System.get");
