+++ lib/static-analysis/StaticEvaluator.js
@@ -201,6 +201,28 @@ const StaticNodeTypeEvaluator = {
 			else if(arg.isNumber()) return new BasicEvaluatedExpression().setString("number").setRange(expression.range);
 			else if(arg.isBoolean()) return new BasicEvaluatedExpression().setString("boolean").setRange(expression.range);
 			else if(arg.isArray() || arg.isConstArray() || arg.isRegExp()) return new BasicEvaluatedExpression().setString("object").setRange(expression.range);
+
+			if(expression.argument.type === "Identifier") {
+				var name = parser.nameInCurrentScope(expression.argument.name);
+				if(parser.scope.definitions.indexOf(name) === -1) {
+					return parser.applyPluginsBailResult1("evaluate typeof " + name, expression);
+				}
+			}
+			if(expression.argument.type === "MemberExpression") {
+				var expressionArgument = expression.argument;
+				var exprName = [];
+				while(expressionArgument.type === "MemberExpression" && !expressionArgument.computed) {
+					exprName.unshift(parser.nameInCurrentScope(expressionArgument.property.name));
+					expressionArgument = expressionArgument.object;
+				}
+				if(expressionArgument.type === "Identifier") {
+					exprName.unshift(parser.nameInCurrentScope(expressionArgument.name));
+					if(parser.scope.definitions.indexOf(name) === -1) {
+						exprName = exprName.join(".");
+						return parser.applyPluginsBailResult1("evaluate typeof " + exprName, expression);
+					}
+				}
+			}
 		} else if(expression.operator === "!") {
 			var argument = parser.evaluateExpression(expression.argument);
 			if(!argument) return;
